requirements.txt
pandas
scikit-learn
tqdm
numpy
faiss-cpu
python-dotenv
